pyproject.toml
[project]
name = "ComfyUI-DownloadFile"
description = "Single ComfyUI node to securely download image/audio files from a URL."
version = "1.0.0"
license = "MIT"
dependencies = [
    "requests>=2.28.2",
    "numpy>=1.24.0",
    "pillow>=9.4.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[project.urls]
Repository = "https://github.com/serious-factory/ComfyUI-DownloadFile"

[tool.comfy]
PublisherId = "serious-factory"
DisplayName = "Download File"
Icon = ""
